+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Change history for ui-courses
 
+## [5.0.0](https://github.com/folio-org/ui-courses/tree/v5.0.0) (IN PROGRESS)
+
+* Upgrade to Stripes v7. Fixes UICR-147.
+
 ## [4.2.1](https://github.com/folio-org/ui-courses/tree/v4.2.1) (2021-06-17)
 
 * Added usage of <NoValue> component for a11y. UICR-100